Fix insufficient synchronization in examples (#30)
The examples use `Arc::strong_count` in order to await all other threads
having terminated. Unfortunately, `Arc::strong_count` uses a `Relaxed`
atomic to load the value, so this can not be used to establish an
ordering, and later leads to a weak memory data race. The alternative
is to use `get_mut`, which uses the internal `is_unique` function, which
is unfortunately private and therefore not usable directly.